feat(webbrowser): add macOS support to build configuration (#2)

+++ librevyv/src/connector.h
@@ -2,6 +2,7 @@
 #define REVYV_CONNECTOR_H
 
 #include "socket.h"
+#include <cstddef>
 #include <cerrno>
 #include <compositor/types.h>
 #include <cstring>
@@ -28,11 +29,11 @@ public:
 
     ~Connector();
 
-    uint32_t window_create(unsigned char* data, uint64_t size, double x, double y, double width, double height, WindowRasterType raster_type);
+    uint32_t window_create(unsigned char* data, size_t size, double x, double y, double width, double height, WindowRasterType raster_type);
 
-    void window_update_pixels(uint32_t window_id, unsigned char* data, uint64_t size, double x, double y, double width, double height);
+    void window_update_pixels(uint32_t window_id, unsigned char* data, size_t size, double x, double y, double width, double height);
 
-    void window_resize(uint32_t window_id, unsigned char* data, uint64_t size, double width, double height);
+    void window_resize(uint32_t window_id, unsigned char* data, size_t size, double width, double height);
 
     void window_change_visiblity(uint32_t window_id, bool visible);
